--- a/plugins/basecontrollers/basecontrollers.cpp
+++ b/plugins/basecontrollers/basecontrollers.cpp
@@ -12,40 +12,8 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#include "plugindefs.h"
-#include <openrave/plugin.h>
+#include "basecontrollers.h"
 
-ControllerBasePtr CreateIdealController(EnvironmentBasePtr penv, std::istream& sinput);
-ControllerBasePtr CreateIdealVelocityController(EnvironmentBasePtr penv, std::istream& sinput);
-ControllerBasePtr CreateRedirectController(EnvironmentBasePtr penv, std::istream& sinput);
-
-InterfaceBasePtr CreateInterfaceValidated(InterfaceType type, const std::string& interfacename, std::istream& sinput, EnvironmentBasePtr penv)
-{
-    switch(type) {
-    case PT_Controller:
-        if( interfacename == "idealcontroller") {
-            return CreateIdealController(penv,sinput);
-        }
-        else if( interfacename == "idealvelocitycontroller") {
-            return CreateIdealVelocityController(penv,sinput);
-        }
-        else if( interfacename == "redirectcontroller" ) {
-            return CreateRedirectController(penv,sinput);
-        }
-        break;
-    default:
-        break;
-    }
-    return InterfaceBasePtr();
-}
-
-void GetPluginAttributesValidated(PLUGININFO& info)
-{
-    info.interfacenames[PT_Controller].push_back("IdealController");
-    info.interfacenames[PT_Controller].push_back("IdealVelocityController");
-    info.interfacenames[PT_Controller].push_back("RedirectController");
-}
-
-OPENRAVE_PLUGIN_API void DestroyPlugin()
-{
+OPENRAVE_PLUGIN_API RavePlugin* CreatePlugin() {
+    return new BaseControllersPlugin();
 }