--- a/plugins/baserobots/baserobots.cpp
+++ b/plugins/baserobots/baserobots.cpp
@@ -14,50 +14,8 @@
 //
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#include "plugindefs.h"
-#include <openrave/plugin.h>
+#include "baserobots.h"
 
-static UserDataPtr s_RegisteredReader;
-
-RobotBasePtr CreateCollisionMapRobot(EnvironmentBasePtr penv, std::istream& sinput);
-RobotBasePtr CreateConveyorRobot(EnvironmentBasePtr penv, std::istream& sinput);
-
-void RegisterCollisionMapRobotReaders(std::list< UserDataPtr >& listRegisteredReaders);
-void RegisterConveyorReaders(std::list< UserDataPtr >& listRegisteredReaders);
-
-static std::list< UserDataPtr >* s_listRegisteredReaders = NULL; ///< have to make it a pointer in order to prevent static object destruction from taking precedence
-
-InterfaceBasePtr CreateInterfaceValidated(InterfaceType type, const std::string& interfacename, std::istream& sinput, EnvironmentBasePtr penv)
-{
-    if( !s_listRegisteredReaders ) {
-        s_listRegisteredReaders = new list< UserDataPtr >();
-        RegisterCollisionMapRobotReaders(*s_listRegisteredReaders);
-        RegisterConveyorReaders(*s_listRegisteredReaders);
-    }
-
-    switch(type) {
-    case PT_Robot:
-        if( interfacename == "collisionmaprobot") {
-            return CreateCollisionMapRobot(penv,sinput);
-        }
-        else if( interfacename == "conveyor" ) {
-            return CreateConveyorRobot(penv,sinput);
-        }
-        break;
-    default:
-        break;
-    }
-    return InterfaceBasePtr();
-}
-
-void GetPluginAttributesValidated(PLUGININFO& info)
-{
-    info.interfacenames[PT_Robot].push_back("CollisionMapRobot");
-    info.interfacenames[PT_Robot].push_back("Conveyor");
-}
-
-OPENRAVE_PLUGIN_API void DestroyPlugin()
-{
-    delete s_listRegisteredReaders;
-    s_listRegisteredReaders = NULL;
+OPENRAVE_PLUGIN_API RavePlugin* CreatePlugin() {
+    return new BaseRobotsPlugin();
 }