This merge request seeks to organize the plugin interface of OpenRAVE. The functions that can be called on a plugin are now collected into a virtual class that all plugins must inherit from and override the pure virtual methods.

The symbols that needs to be loaded from a plugin shared object is reduced to just a single C call - `RavePlugin* CreatePlugin();`, eliminating the need to check for the existence of multiple symbols. Previous implementation required a thread to lazy load libraries as checking for multiple symbols was time consuming; however it still eager-loaded the `OpenRaveGetPluginAttributes` function. Here since there's only one function to load, and we already have the time budget to eager-load one function, there is no longer a need for the extra thread, and all the supporting mutexes and condition_variables surrounding that thread.

The previous `RegisteredInterface` object has also been folded into the plugin object class by recognizing it as a special subclass of plugins that don't inhabit a separate binary. This eliminates the need for a different set of code paths to handle them - just create a `VirtualPlugin` object and set the fields to what was previously being set in `RegisteredInterface`.

All the plugins have been updated to reflect the changes made to the `openrave/plugin.h` header. Some plugins have interfaces that don't have an implementation; those interfaces have been commented out (linking fails otherwise).

The advantages of this merge request is that plugin and interface creation code are now neatly segregated into a class, and statically compiling in plugins will not require large amounts of code duplication.

I recognize that this is a rather significant change that breaks existing conventions. Effort has been made to ensure that internal libraries are also updated, tested, and fixed in sync. I may have some unknown unknowns, kindly review.