@rdiankov 2022-10-25
@undisputed-seraphim This way might be slower now since dlopen has to process more symbols in the SO.  Please do some perf experiments with starting the plugin database and getting all plugin info comparing your way and the current production way.
thanks