@undisputed-seraphim 2022-11-10
I timed the changes in two ways. For every test, OpenRave was able to find 14 plugins, the ones bundled in this repository as well as elsewhere in our system.

1. Using `time openrave --listplugins`. Ran it for 5 times each, and the final column is the average.

Before
```
real(s) |0.056 |0.063 |0.055 |0.085 |0.062 | 0.0642
user(s) |0.030 |0.038 |0.027 |0.044 |0.027 | 0.0332
sys(s)  |0.018 |0.018 |0.020 |0.033 |0.027 | 0.0232
```
After
```
real(s) |0.070 |0.065 |0.069 |0.065 |0.070 | 0.0678
user(s) |0.039 |0.042 |0.037 |0.042 |0.037 | 0.0394
sys(s)  |0.023 |0.016 |0.025 |0.016 |0.026 | 0.0212
```
New changes are slower by about 0.002 to 0.006 seconds.

2. Using `high_resolution_clock` to time the execution of running `RaveInit` and `RaveCreateEnvironment`. Ran each test 3 times, final column is the average.

Times are in nanoseconds.

Before
```
1130700
1120776
1097157
Average: 1116211
```
After
```
1115112
1111894
1114423
Average: 1113810 (-2401 nanoseconds)
```
New changes are faster by 2401 nanoseconds (0.002401 milliseconds)

I contend that this is negligible and furthermore only performed once per invocation of `RaveCreateEnvironment`.

On the other hand, this changes brings us some benefits:
- As the plugin is now represented by an object, we can use RAII to correctly and safely release any memory (or other resource) required by the plugin.
- As the lifetimes of values and variables is now controlled by the object, we don't have to use hacky null-checks and workarounds for static values [such as these](https://github.com/rdiankov/openrave/blob/master/plugins/oderave/oderave.cpp#L23).
- Due to the above point, we can avoid [static initialization fiasco](https://en.cppreference.com/w/cpp/language/siof), as now it is controlled by the object in a well-defined way.
- Compilation will fail if required functions are not implemented - as opposed to failing only at runtime, and requiring a rebuild.
- The plugin entrypoint is now represented by a single C-compatible linkage, and is therefore cross-platform and also language agnostic if anyone wants to create bindings.
- New `libopenrave0.101.so` binary is slightly smaller (old: 4235680 bytes, new: 4223392 bytes; smaller by exactly 12 kilobytes!)
- It enables us to implement a fully static OpenRAVE library with all plugins embedded in it for more performance gains.