@cielavenir 2022-11-10
I'm personally interested in COM (where all interfaces inherit IUnknown https://github.com/jinfeihan57/p7zip/blob/master/CPP/Common/MyWindows.h#L119)