@cielavenir 2022-11-10
In this way the only exported symbol is "CreateObject", so it costs less for ld-linux