@rdiankov 2022-11-14
@undisputed-seraphim Can you send the c++ code you used to do perf tests? thanks!