@undisputed-seraphim 2022-11-18
> @undisputed-seraphim Can you send the c++ code you used to do perf tests? thanks!

For temporary testing, I merged the branch `liboon/tests`, and built and executed the tests.
The test executable is named `openrave_tests`

 `liboon/tests` is a branch I am using to develop a set of tests written in C++.