@undisputed-seraphim 2022-11-21
@rdiankov I've written another benchmark that produces a more stark difference.

```c++
#include <chrono>
#include <iostream>
#include <vector>
#include <openrave/openrave.h>
#include <openrave-core.h>

constexpr int ITERATIONS = 1000;

int main() {
    std::vector<std::chrono::nanoseconds> deltas = {};
    deltas.resize(ITERATIONS);
    for (int i = 0; i < ITERATIONS; ++i) {
        auto start = std::chrono::high_resolution_clock::now();
        OpenRAVE::RaveInitialize(true);
        OpenRAVE::RaveDestroy();
        deltas[i] = (std::chrono::high_resolution_clock::now() - start);
    }
    uint64_t average_ns = 0;
    for (const auto& delta : deltas) {
        average_ns += delta.count();
    }
    average_ns /= ITERATIONS;
    double average_ms = average_ns / 1'000'000.0;
    std::cout << average_ns << " nanoseconds (" << average_ms << " milliseconds), on average, across 1000 iterations" << std::endl;
    return 0;
}
```

Linked against current production:
807368 nanoseconds (0.807368 milliseconds) on average across 1000 iterations.
Linked against my changes:
244185 nanoseconds (0.244185 milliseconds), on average across 1000 iterations.
Indicating a ~3.5 times speedup.