@rdiankov 2022-11-21
@undisputed-seraphim For the c++ benchmark, please add getting the list of plugin information. thanks