@undisputed-seraphim 2022-11-22
@rdiankov 

```c++
#include <chrono>
#include <iostream>
#include <array>
#include <openrave/openrave.h>
#include <openrave-core.h>

constexpr int ITERATIONS = 1000;

int main() {
    std::list<std::pair<std::string, OpenRAVE::PLUGININFO>> plugins;
    std::array<std::chrono::nanoseconds, ITERATIONS> deltas;
    for (int i = 0; i < ITERATIONS; ++i) {
        plugins.clear();
        auto start = std::chrono::high_resolution_clock::now();
        OpenRAVE::RaveInitialize(true);
        OpenRAVE::RaveGetPluginInfo(plugins);
        OpenRAVE::RaveDestroy();
        deltas[i] = (std::chrono::high_resolution_clock::now() - start);
    }
    std::cout << "Retrieved information about " << plugins.size() << " plugins." << std::endl;
    uint64_t average_ns = 0;
    for (const auto& delta : deltas) {
        average_ns += delta.count();
    }
    average_ns /= ITERATIONS;
    double average_ms = average_ns / 1'000'000.0;
    std::cout << average_ns << " nanoseconds (" << average_ms << " milliseconds), on average, across 1000 iterations" << std::endl;
    return 0;
}
```

Output (production):
```
Retrieved information about 15 plugins.
783248 nanoseconds (0.783248 milliseconds), on average, across 1000 iterations
```

Output (newPlugin):
```
Retrieved information about 15 plugins.
251012 nanoseconds (0.251012 milliseconds), on average, across 1000 iterations
```