@rdiankov 2022-12-04
@undisputed-seraphim Awesome, good job! So once I merge this, we need to also change all our internal plugins in controllercommon and planningcommon right? Can you prepare merge requests for that?
Thanks!